I was wrong; name bundles by hash of contents
A bundle is not identified by its "access module" because it may
have more than one. Bundles are now named by taking all contained
module paths, and hashing their concatenation.

As a result, if a root is added to the set that depends on a bundle,
B, but the contents of B stay the same, the name of B will stay
the same, and browsers will not have to redownload it.